Readme: enhance PHPCS example
When the `-q` option is used for PHPCS, no output will be shown in the action logs anymore, other than the exit code.

At times, it may be helpful to still see output in the logs as well (to, for instance, verify the step is running correctly).

This additional example shows how to get that working.

Signed-off-by: jrfnl <jrfnl@users.noreply.github.com>